made settings work using maps

+++ lib/pages/splash_screen/view/splash_screen.dart
@@ -40,11 +40,13 @@ class _SplashScreenState extends State<SplashScreen> {
             user_data['username'],
             user_data['password'],
             tokens['access_token']);
-        if (spotify_data['data']['success'] == true) {
-          if (spotify_data['data']['code'] != 'not_playing') {
-            spotify_data = spotify_data['data']['data'];
-            await Storage.deleteData('playing_song');
-            await Storage.saveData(spotify_data, 'playing_song');
+        if (spotify_data['data'] != null) {
+          if (spotify_data['data']['success'] == true) {
+            if (spotify_data['data']['code'] != 'not_playing') {
+              spotify_data = spotify_data['data']['data'];
+              await Storage.deleteData('playing_song');
+              await Storage.saveData(spotify_data, 'playing_song');
+            }
           }
         }
